refactor(svar-2): move process_chromosome into orchestrator.rs; slim lib.rs
lib.rs is now just the pyo3 _core surface + run_conversion_pipeline. The
long-allele writer thread body moves into writer::run_long_allele_writer.
